+++ packages/add-component/template/component/src/__name__.vue
@@ -1,9 +1,8 @@
 <script lang="ts" setup name="F<%= displayName %>">
-  import { Props, Emits } from './props'
+  import { Props } from './props'
   import type { <%= displayName %>PropsType } from './interface'
 
   const prop: <%= displayName %>PropsType = defineProps(Props)
-  defineEmits(Emits)
 </script>
 
 <template>